p03: Add file upload/download routes

+++ server-spring/src/main/java/com/nogenem/skyapp/model/Attachment.java
@@ -9,7 +9,7 @@ public class Attachment {
 
   private String originalName;
 
-  private int size;
+  private long size;
 
   private String path;
